+++ src/main/java/efub/agoda_server/security/oauth2/CustomOAuth2User.java
@@ -1,6 +1,6 @@
 package efub.agoda_server.security.oauth2;
 
-import efub.agoda_server.domain.entity.User;
+import efub.agoda_server.user.domain.User;
 import lombok.Getter;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.oauth2.core.user.OAuth2User;
